Move shared types into dedicated files
