+++ modules/graphql-router/src/config/constants.ts
@@ -1,4 +1,8 @@
-import { featureFlagDefaults, setsProperties } from '@overture-stack/arranger-types/configs';
+import {
+	configArrangerNetworkProperties,
+	featureFlagDefaults,
+	setsProperties,
+} from '@overture-stack/arranger-types/configs';
 import {
 	chartsProperties,
 	downloadProperties,
@@ -37,7 +41,7 @@ const fallbackCatalogConfigs = {
 		[facetsProperties.AGGS]: [],
 	},
 	[configRootProperties.MATCHBOX]: [],
-	[configRootProperties.NETWORK_AGGREGATION]: [],
+	[configRootProperties.NETWORK_AGGREGATION]: { [configArrangerNetworkProperties.REMOTE_NODES]: [] },
 	[configRootProperties.TABLE]: {
 		[tableProperties.COLUMNS]: [],
 		[tableProperties.MAX_RESULTS_WINDOW]: MAX_RESULTS_WINDOW,
